Learned how to restrict pages

+++ base/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
@@ -12,6 +14,10 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -35,7 +41,6 @@ def logoutUser(request):
     logout(request)
     return redirect('home')
     
-
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -53,30 +58,46 @@ def room(request,pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
+
     if request.method == "POST":
         form = RoomForm(request.POST)
+
         if form.is_valid():
             form.save()
             return redirect('home')
+
     context = {'form': form}
     return render(request, 'base/room_form.html', context) 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed herte!')
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
             return redirect('home')
+
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed herte!')
+
     if request.method == "POST":
         room.delete()
         return redirect('home')
+
     return render(request, 'base/delete.html', {'obj': room})
